requirements.txt
chromadb-client
chromadb
python-dotenv
openai